+++ frontend/lib/screens/adm_page/home_adm_view/home_adm_view.dart
@@ -5,8 +5,6 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:research_repository/research_repository.dart';
 import 'package:user_repository/user_repository.dart';
 
-import 'widgets/search_tile.dart';
-
 class HomeAdmView extends StatefulWidget {
   const HomeAdmView({super.key});
 
@@ -132,6 +130,51 @@ class _HomeAdmViewState extends State<HomeAdmView> {
           },
         );
       },
+    );
+  }
+}
+
+// widget for discovered searches
+class SearchTile extends StatefulWidget {
+  final String name;
+  final String status;
+  const SearchTile({super.key, required this.name, required this.status});
+
+  @override
+  State<SearchTile> createState() => _SearchTileState();
+}
+
+class _SearchTileState extends State<SearchTile> {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Align(
+          alignment: Alignment.topLeft,
+          child: Text(
+            widget.name,
+            style: TextStyle(fontSize: 14.sp, color: Colors.white),
+          ),
+        ),
+        SizedBox(height: 1.h),
+        Row(
+          children: [
+            Container(
+              width: 4.w,
+              height: 4.h,
+              decoration: BoxDecoration(
+                color: widget.status == 'OPEN' ? Colors.green : Colors.red,
+                shape: BoxShape.circle,
+              ),
+            ),
+            Text(
+              ' ${widget.status} | 1246 entradas',
+              style: TextStyle(fontSize: 10.sp, color: Colors.white),
+            ),
+          ],
+        ),
+        SizedBox(height: 20.h),
+      ],
     );
   }
 }
