--- frontend/lib/utils/repository_utils.dart
@@ -1,3 +1,0 @@
-class SpringConection {
-  final String adressIP = 'http://192.168.15.12:8080';
-}
